Remove game on mobile
It worked but it was just clunky. My personal phone couldn't maintain a good frame rate, which cause janky collision detection. Also while tapping and dragging to move paddle, it was easy to start scrolling which caused address bar on mobile to unfold. Generally just not a good enough experience on mobile to justify keeping it.